First buildable version
Changed LevelTrigger so it doesn't use Unity Edditor tools
Changed speed of enemies so it is not dependent on Tickspeed.
Changed Item Hooking so it is not dependent on TickSpeed

+++ DH2650/Assets/Scripts/HealthBar.cs
@@ -11,14 +11,14 @@ public class HealthBar : MonoBehaviour
 
     public PlayerHealth playerHealth;
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
 
         text.text = health.ToString() + "/" + playerHealth.maxHealth.ToString();
     }
 
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         slider.maxValue = health;
         slider.value = health;
